WIP home refactoring

+++ core/data/src/main/java/com/core/data/model/dto/TokenBalanceDto.kt
@@ -23,7 +23,7 @@ fun TokenBalanceDto.asEntity(chainId: Int): TokenBalanceEntity {
     return TokenBalanceEntity(
         contractAddress = contractAddress,
         chainId = chainId,
-        tokenBalance = tokenBalance.toInt()
+        tokenBalance = tokenBalance.toDouble()
 
     )
 }
